requirements.txt
torch
torchvision
torchaudio
tensorflow
tensorflow-datasets
numpy
pandas
scikit-learn
matplotlib
seaborn
plotly
opencv-python
Pillow
streamlit
scipy
jupyter


